# docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Tarot Bot API",
    "description": "Everything you need for digital Tarot readings"
  },
  "host": "tarot-bot-api.vercel.app",
  "basePath": "/",
  "schemes": [],
  "paths": {
    "/random": {
      "get": {
        "description": "Returns a random card",
        "parameters": [
          {
            "name": "deck",
            "description": "Bit mask or string representation of subset of deck. Used to limit deck to major or minor arcana only, or remove cards from the deck as they’re drawn. You can use https://api.tarotbot.cards/calculator to calculate the deck state",
            "in": "query",
            "type": "string"
          },
          {
            "name": "reverseChance",
            "description": "Percent chance the card will be reversed. Default 0.5",
            "in": "query",
            "type": "string"
          },
          {
            "name": "images",
            "description": "The (encoded uri component) url of a formatted json file with urls or string keyword of a preexisting deck for card images and descriptions. rws, rider-waite, rider-waite-smith return Rider Waite Smith deck; tarotbot, tb return Tarot Bot deck. Default rws",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/daily": {
      "get": {
        "description": "Returns a unique Tarot Card for the given date",
        "parameters": [
          {
            "name": "date",
            "description": "The day the returned card will represent in MMDDYYYY format. Default today",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/archive": {
      "get": {
        "description": "Returns past Daily Readings",
        "parameters": [
          {
            "name": "start",
            "description": "The date to start counting back from, combines with offset. Default today",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "description": "The maximum number of daily readings to return. Default 100, max 10000",
            "in": "query",
            "type": "string"
          },
          {
            "name": "offset",
            "description": "The number of days in the past from the start date to count back from. Default 0",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/spread": {
      "get": {
        "description": "Returns multiple random cards in the defined arrangement. EXPERIMENTAL (please be gentle)",
        "parameters": [
          {
            "name": "format",
            "description": "The numerical dimensions (ie 900,530) and card positions (ie 300,0) describing the arrangement or a string keyword of a preexisting arrangement. Note: a card’s default dimensions are 300x530. Example: pastpresentfuture is identical to 900,530,0,0,300,0,600,0. Default pastpresentfuture",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/card": {
      "get": {
        "description": "Returns a specific card",
        "parameters": [
          {
            "name": "name",
            "description": "String representation of the card",
            "in": "query",
            "type": "string"
          },
          {
            "name": "images",
            "description": "The (encoded uri component) url of a formatted json file with urls for card images. Used to change the images that tarot bot returns",
            "in": "query",
            "type": "string"
          },
          {
            "name": "reversed",
            "description": "Return card as reversed. Default false",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cards": {
      "get": {
        "description": "Returns all cards",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}